+++ app/src/index.html
@@ -495,7 +495,7 @@
         </div>
         <div class="separator"></div>
         <div class="operations single-choice">
-          <button id="operation-system-move" title="Move" class="selected">
+          <button id="operation-move" title="Move" class="selected">
             <img src="./assets/icon-move.svg" />
             <div class="keyboard">1</div>
           </button>
@@ -647,6 +647,10 @@
               <td>Zoom out of the canvas.</td>
             </tr>
             <tr>
+              <td><img src="./assets/icon-move.svg" /></td>
+              <td>Move a system or a link in the canvas.</td>
+            </tr>
+            <tr>
               <td><img src="./assets/icon-eraser.svg" /></td>
               <td>Remove a system or a link from the canvas.</td>
             </tr>
@@ -657,10 +661,6 @@
             Operations on systems
           </caption>
           <tbody>
-            <tr>
-              <td><img src="./assets/icon-move.svg" /></td>
-              <td>Move a system in the canvas.</td>
-            </tr>
             <tr>
               <td><img src="./assets/icon-system.svg" /></td>
               <td>Add a system to the canvas.</td>
